Add base mutator
