+++ src/common/storages/fuse/src/operations/append.rs
@@ -22,9 +22,9 @@ use common_exception::Result;
 use common_pipeline_core::processors::port::InputPort;
 use common_pipeline_core::Pipeline;
 use common_pipeline_core::SinkPipeBuilder;
-use common_pipeline_transforms::processors::transforms::transform_compact::TransformCompact;
 use common_pipeline_transforms::processors::transforms::BlockCompactor;
 use common_pipeline_transforms::processors::transforms::ExpressionTransform;
+use common_pipeline_transforms::processors::transforms::TransformCompact;
 use common_pipeline_transforms::processors::transforms::TransformSortPartial;
 use common_pipeline_transforms::processors::ExpressionExecutor;
 use common_planners::Expression;
