+++ src/common/storages/fuse/src/operations/mutation/deletion_mutator.rs
@@ -12,28 +12,18 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-use std::collections::hash_map::RandomState;
-use std::collections::HashMap;
 use std::sync::Arc;
 
 use common_catalog::table_context::TableContext;
 use common_datablocks::DataBlock;
-use common_exception::ErrorCode;
 use common_exception::Result;
-use common_fuse_meta::meta::BlockMeta;
 use common_fuse_meta::meta::ClusterStatistics;
 use common_fuse_meta::meta::Location;
-use common_fuse_meta::meta::SegmentInfo;
 use common_fuse_meta::meta::TableSnapshot;
-use opendal::Operator;
 
-use crate::io::write_meta;
 use crate::io::BlockWriter;
-use crate::io::MetaReaders;
-use crate::io::SegmentWriter;
 use crate::io::TableMetaLocationGenerator;
-use crate::statistics::reducers::reduce_block_metas;
-use crate::statistics::reducers::reduce_statistics;
+use crate::operations::mutation::BaseMutator;
 use crate::statistics::ClusterStatsGenerator;
 
 pub enum Deletion {
@@ -41,19 +31,8 @@ pub enum Deletion {
     Remains(DataBlock),
 }
 
-pub struct Replacement {
-    original_block_loc: Location,
-    new_block_meta: Option<BlockMeta>,
-}
-
-pub type SegmentIndex = usize;
-
 pub struct DeletionMutator<'a> {
-    mutations: HashMap<SegmentIndex, Vec<Replacement>>,
-    ctx: &'a Arc<dyn TableContext>,
-    location_generator: &'a TableMetaLocationGenerator,
-    base_snapshot: &'a TableSnapshot,
-    data_accessor: Operator,
+    base_mutator: BaseMutator<'a>,
     cluster_stats_gen: ClusterStatsGenerator,
 }
 
@@ -64,116 +43,16 @@ impl<'a> DeletionMutator<'a> {
         base_snapshot: &'a TableSnapshot,
         cluster_stats_gen: ClusterStatsGenerator,
     ) -> Result<Self> {
-        let data_accessor = ctx.get_storage_operator()?;
+        let base_mutator = BaseMutator::try_create(ctx, location_generator, base_snapshot)?;
         Ok(Self {
-            mutations: HashMap::new(),
-            ctx,
-            location_generator,
-            base_snapshot,
-            data_accessor,
+            base_mutator,
             cluster_stats_gen,
         })
     }
 
     pub async fn into_new_snapshot(self) -> Result<(TableSnapshot, String)> {
-        let snapshot = self.base_snapshot;
-        let mut new_snapshot = TableSnapshot::from_previous(snapshot);
-
-        // takes away the segments, they are being mutated
-        let mut segments_editor = HashMap::<_, _, RandomState>::from_iter(
-            std::mem::take(&mut new_snapshot.segments)
-                .into_iter()
-                .enumerate(),
-        );
-
-        let segment_reader = MetaReaders::segment_info_reader(self.ctx.as_ref());
-
-        let segment_info_cache = self
-            .ctx
-            .get_storage_cache_manager()
-            .get_table_segment_cache();
-        let seg_writer = SegmentWriter::new(
-            &self.data_accessor,
-            self.location_generator,
-            &segment_info_cache,
-        );
-
-        // apply mutations
-        for (seg_idx, replacements) in self.mutations {
-            let segment = {
-                let (path, version) = &snapshot.segments[seg_idx];
-                segment_reader.read(&path, None, *version).await?
-            };
-
-            // collects the block locations of the segment being modified
-            let block_positions = segment
-                .blocks
-                .iter()
-                .enumerate()
-                .map(|(idx, meta)| (&meta.location, idx))
-                .collect::<HashMap<_, _>>();
-
-            // prepare the new segment
-            let mut new_segment = SegmentInfo::new(segment.blocks.clone(), segment.summary.clone());
-
-            // take away the blocks, they are being mutated
-            let mut block_editor = HashMap::<_, _, RandomState>::from_iter(
-                std::mem::take(&mut new_segment.blocks)
-                    .into_iter()
-                    .enumerate(),
-            );
-
-            for replacement in replacements {
-                let position = block_positions
-                    .get(&replacement.original_block_loc)
-                    .ok_or_else(|| {
-                        ErrorCode::LogicalError(format!(
-                            "block location not found {:?}",
-                            &replacement.original_block_loc
-                        ))
-                    })?;
-                if let Some(block_meta) = replacement.new_block_meta {
-                    block_editor.insert(*position, block_meta);
-                } else {
-                    block_editor.remove(position);
-                }
-            }
-            // assign back the mutated blocks to segment
-            new_segment.blocks = block_editor.into_values().collect();
-
-            if new_segment.blocks.is_empty() {
-                // remove the segment if no blocks there
-                segments_editor.remove(&seg_idx);
-            } else {
-                // re-calculate the segment statistics
-                let new_summary = reduce_block_metas(&new_segment.blocks)?;
-                new_segment.summary = new_summary;
-                // write down new segment
-                let new_segment_location = seg_writer.write_segment(new_segment).await?;
-                segments_editor.insert(seg_idx, new_segment_location);
-            }
-        }
-
-        // assign back the mutated segments to snapshot
-        new_snapshot.segments = segments_editor.into_values().collect();
-
-        let mut new_segment_summaries = Vec::with_capacity(new_snapshot.segments.len());
-        for (loc, ver) in &new_snapshot.segments {
-            let seg = segment_reader.read(loc, None, *ver).await?;
-            new_segment_summaries.push(seg.summary.clone())
-        }
-
-        // update the summary of new snapshot
-        let new_summary = reduce_statistics(&new_segment_summaries)?;
-        new_snapshot.summary = new_summary;
-
-        // write down the new snapshot
-        let snapshot_loc = self.location_generator.snapshot_location_from_uuid(
-            &new_snapshot.snapshot_id,
-            new_snapshot.format_version(),
-        )?;
-        write_meta(&self.data_accessor, &snapshot_loc, &new_snapshot).await?;
-        Ok((new_snapshot, snapshot_loc))
+        let (segments, summary) = self.base_mutator.generate_segments().await?;
+        self.base_mutator.into_new_snapshot(segments, summary).await
     }
 
     /// Records the replacements:  
@@ -189,21 +68,19 @@ impl<'a> DeletionMutator<'a> {
         let new_block_meta = if replace_with.num_rows() == 0 {
             None
         } else {
-            let block_writer =
-                BlockWriter::new(self.ctx, &self.data_accessor, self.location_generator);
+            let block_writer = BlockWriter::new(
+                self.base_mutator.ctx,
+                &self.base_mutator.data_accessor,
+                self.base_mutator.location_generator,
+            );
             let cluster_stats = self
                 .cluster_stats_gen
                 .gen_with_origin_stats(&replace_with, origin_stats)?;
             Some(block_writer.write(replace_with, cluster_stats).await?)
         };
         let original_block_loc = location_of_block_to_be_replaced;
-        self.mutations
-            .entry(seg_idx)
-            .or_default()
-            .push(Replacement {
-                original_block_loc,
-                new_block_meta,
-            });
+        self.base_mutator
+            .add_mutation(seg_idx, original_block_loc, new_block_meta);
         Ok(())
     }
 }
